Reduce Large Files Step 4: extract headless command families
## Summary

This slice reorganizes the headless command surface in `packages/app/src`.

The single `headless.ts` entry now delegates each command family to its own handler module behind one dispatcher.

Observable CLI behavior stays the same.

<details>
<summary>Review metadata</summary>

Review Claim: Approve grouping headless commands into per-family handler modules behind one CLI dispatcher, with no behavior change.

Review Lane: refactor

Review Unit: activation-surface

Safety Invariant: Each handler keeps the same flags, exit codes, and delegation paths, so the CLI surface is unchanged and reviewable in isolation.

Slice Rationale: Grouping by headless command family keeps each handler module reviewable on its own.

</details>

## Non-goals

- No behavior change to any headless command; behavior unchanged.
- No change to task-runner or persistence modules.

## Test Plan

- [ ] `cd packages/app && pnpm test`

## Revert Plan

- Safe to revert? Yes
- Revert command: `git revert <sha>`
- Post-revert steps: None
- Data migration? No
