+++ packages/app/src/headless.ts
@@ -1,71 +1,89 @@
 /**
  * Headless CLI logic extracted from main.ts.
  *
- * All functions that implement `--headless <command>` live here.
- * They receive shared services via a `HeadlessDeps` object instead of
- * accessing module-level variables directly.
+ * This file owns the headless command router (`runHeadless`), the usage text,
+ * the `set <sub>` configure family, and the long-running `slack`/`worker`
+ * commands. The per-family command handlers live in dedicated modules:
  *
- * Business logic (orchestrator mutations) lives in workflow-actions.ts.
- * This file handles CLI parsing, TaskRunner lifecycle, and output formatting.
+ *   - `headless-shared.ts`        — shared infra (HeadlessDeps, TaskRunner
+ *                                    wiring, workflow tracking, restore /
+ *                                    preemption / flag-parsing helpers)
+ *   - `headless-run-resume.ts`    — run / resume / watch / retry / recreate /
+ *                                    rebase / fork / fix / resolve-conflict
+ *   - `headless-query-list.ts`    — read-only `query <sub>` + cost + session
+ *   - `headless-approve-delete.ts`— approve / reject / input / select +
+ *                                    cancel / delete / detach / open-terminal
+ *
+ * The public surface (consumed by main.ts, IPC, and tests) is re-exported
+ * below so import sites are unchanged. Business logic (orchestrator
+ * mutations) lives in workflow-actions.ts.
  */
 
-import type { BundledSkillsInstallMode, BundledSkillsStatus, Logger } from '@invoker/contracts';
+import type { BundledSkillsInstallMode } from '@invoker/contracts';
 import { makeEnvelope } from '@invoker/contracts';
-import type { AgentSessionData } from '@invoker/contracts';
-import { OrchestratorErrorCode } from '@invoker/workflow-core';
-import type { Attempt, Orchestrator, CommandService, TaskDelta, TaskState } from '@invoker/workflow-core';
-import type { SQLiteAdapter } from '@invoker/data-store';
-import { Channels } from '@invoker/transport';
-import type { MessageBus } from '@invoker/transport';
-import {
-  ExecutorRegistry,
-  TaskRunner,
-  GitHubMergeGateProvider,
-  ReviewProviderRegistry,
-  remoteFetchForPool,
-  registerBuiltinAgents,
-  assertPlanExecutionAgentsRegistered,
-  type AgentRegistry,
-  type TaskHeartbeatEvent,
-} from '@invoker/execution-engine';
-import { loadConfig, resolveSecretsFilePath, type InvokerConfig } from './config.js';
-import { backupPlan } from './plan-backup.js';
+import type { TaskState } from '@invoker/workflow-core';
+import type { TaskRunner } from '@invoker/execution-engine';
 import { startApiServer } from './api-server.js';
-import { WorkflowMutationFacade } from './workflow-mutation-facade.js';
 import {
   parseMetadataValue,
   setTaskMetadata,
   setWorkflowMetadata,
 } from './metadata-setter.js';
 import {
-  approveTask,
-  autoFixOnReviewGateFailure,
   deleteAllWorkflows as sharedDeleteAllWorkflows,
-  fixWithAgentAction,
-  rebaseRetry,
-  rebaseRecreate,
-  resolveConflictAction,
-  forkWorkflow as sharedForkWorkflow,
   setWorkflowMergeMode,
 } from './workflow-actions.js';
 import { normalizeMergeModeForPersistence } from './merge-mode.js';
-import { parseHeadlessFixArgs } from './auto-fix-intents.js';
-import type { CostGroupDimension } from './cost-rollup.js';
-import { openExternalTerminalForTask } from './open-terminal-for-task.js';
-import {
-  dispatchStartedTasksWithGlobalTopup,
-  executeGlobalTopup,
-  finalizeMutationWithGlobalTopup,
-  isDispatchableLaunch,
-} from './global-topup.js';
+import { isDispatchableLaunch } from './global-topup.js';
 import { LaunchDispatcher } from './launch-dispatcher.js';
 import { createAutoFixRecoveryTick, RECOVERY_WORKER_KIND } from './workers/auto-fix-recovery.js';
-import { resolveHeadlessTargetWorkflowId } from './headless-command-classification.js';
 import { formatHeadlessSetSubcommands } from './headless-command-registry.js';
-import { trackWorkflow } from './headless-watch.js';
-import { preemptWorkflowBeforeMutation, type WorkflowCancelResult } from './workflow-preemption.js';
-import type { WorkflowMutationTiming } from './workflow-mutation-timing.js';
-import type { RuntimeServices } from '@invoker/runtime-service';
+import {
+  type HeadlessDeps,
+  BOLD,
+  RESET,
+  YELLOW,
+  createHeadlessExecutor,
+  wireHeadlessApproveHook,
+  wireHeadlessAutoFix,
+  buildHeadlessApiServerDeps,
+  trackHeadlessWorkflow,
+  restoreWorkflowForTask,
+  restoreWorkflowForTaskUnlessDeleteAllWon,
+  withRestoredTaskUnlessDeleteAllWon,
+} from './headless-shared.js';
+import {
+  headlessRun,
+  headlessResume,
+  headlessWatch,
+  headlessRetryWorkflow,
+  headlessRetryTask,
+  headlessRecreateWorkflow,
+  headlessRecreateTask,
+  headlessRecreateDownstream,
+  headlessForkWorkflow,
+  headlessRebaseRetry,
+  headlessRebaseRecreate,
+  headlessFix,
+  headlessResolveConflict,
+} from './headless-run-resume.js';
+import {
+  headlessQuery,
+  headlessQuerySelect,
+} from './headless-query-list.js';
+import {
+  headlessApprove,
+  headlessReject,
+  headlessInput,
+  headlessSelect,
+  headlessCancel,
+  headlessCancelWorkflow,
+  headlessDeleteWorkflow,
+  headlessDetachWorkflow,
+  headlessOpenTerminal,
+} from './headless-approve-delete.js';
+
+// ── Public re-exports (stable surface for main.ts, IPC, and tests) ──
 
 export {
   DEFAULT_DELEGATION_TIMEOUT_MS,
@@ -80,196 +98,27 @@ export {
 } from './headless-delegation.js';
 export type { DelegationOutcome } from './headless-delegation.js';
 
-// ── HeadlessDeps interface ───────────────────────────────────
+export { createHeadlessExecutor, wireHeadlessAutoFix, wireHeadlessApproveHook };
+export { parseQueryFlags } from './headless-shared.js';
+export type { HeadlessDeps, QueryFlags, HeadlessAutoFixController } from './headless-shared.js';
+export { resolveAgentSession } from './headless-query-list.js';
 
-export interface HeadlessDeps {
-  logger: Logger;
-  orchestrator: Orchestrator;
-  persistence: SQLiteAdapter;
-  executorRegistry: ExecutorRegistry;
-  messageBus: MessageBus;
-  commandService: CommandService;
-  repoRoot: string;
-  invokerConfig: InvokerConfig;
-  initServices: () => Promise<void>;
-  executionAgentRegistry?: AgentRegistry;
-  wireSlackBot: (deps: {
-    executor: TaskRunner;
-    logFn: (source: string, level: string, message: string) => void;
-    approveTaskAction?: (taskId: string) => Promise<void>;
-    onPlanLoaded?: () => void;
-  }) => Promise<any>;
-  getUiPerfStats?: () => Record<string, unknown>;
-  resetUiPerfStats?: () => void;
-  deferRunnableTasks?: (tasks: TaskState[], workflowId?: string) => void;
-  preemptTaskSubgraph?: (taskId: string) => Promise<void>;
-  preemptWorkflowExecution?: (workflowId: string) => Promise<WorkflowCancelResult>;
-  cancelTask?: (taskId: string) => Promise<{ cancelled: string[]; runningCancelled: string[] }>;
-  cancelWorkflow?: (workflowId: string) => Promise<{ cancelled: string[]; runningCancelled: string[] }>;
-  waitForApproval?: boolean;
-  noTrack?: boolean;
-  isStandaloneOwnerIdle?: () => boolean;
-  getBundledSkillsStatus?: () => BundledSkillsStatus;
-  installBundledSkills?: (mode?: BundledSkillsInstallMode) => BundledSkillsStatus;
-  /** Abort signal from the workflow mutation coordinator, if running inside a coordinated mutation. */
-  signal?: AbortSignal;
-  mutationTiming?: WorkflowMutationTiming;
-  runtimeServices?: RuntimeServices;
-  /**
-   * CB.7: provider for the owner's long-lived TaskRunner. When the
-   * launch-outbox is `'active'`, `createHeadlessExecutor` reuses this
-   * instance instead of constructing a fresh `TaskRunner` per command.
-   * Returning `null` (or omitting the provider entirely) falls back to
-   * the legacy behaviour (a new TaskRunner each call). This eliminates
-   * Issue 6 (multi-TaskRunner blindness — each runner has its own
-   * `launchingAttemptIds` Set) once the outbox dispatcher is the only
-   * launch path. The fallback also keeps the function safe to call in
-   * environments without an owner-mode TaskRunner (peer mode, tests).
-   */
-  ownerTaskRunnerProvider?: () => TaskRunner | null;
+// ── Deprecation Warning ─────────────────────────────────────
+
+function warnDeprecated(oldCmd: string, newCmd: string): void {
+  process.stderr.write(
+    `${YELLOW}[deprecated]${RESET} "${oldCmd}" is deprecated. Use "${newCmd}" instead.\n`,
+  );
 }
 
-// ── ANSI Helpers ─────────────────────────────────────────────
-
-const RESET = '\x1b[0m';
-const BOLD = '\x1b[1m';
-const YELLOW = '\x1b[33m';
-
-// ── Shared Helpers ───────────────────────────────────────────
-
-function headlessHeartbeat(
-  taskId: string,
-  event: TaskHeartbeatEvent,
-  deps: Pick<HeadlessDeps, 'orchestrator'>,
-): void {
-  deps.orchestrator.recordTaskHeartbeat(taskId, { at: event.at, source: event.source });
+function assertDeleteAllEnabled(): void {
+  if (process.env.INVOKER_ALLOW_DELETE_ALL === '1') return;
+  throw new Error(
+    'delete-all is disabled by default. Set INVOKER_ALLOW_DELETE_ALL=1 to enable it explicitly.',
+  );
 }
 
-function buildHeadlessApiServerDeps(
-  deps: HeadlessDeps,
-  taskExecutor: TaskRunner,
-): { mutations: WorkflowMutationFacade; deleteWorkflow: (id: string) => Promise<void>; detachWorkflow: (id: string, upstreamId: string) => Promise<void> } {
-  return {
-    mutations: new WorkflowMutationFacade({
-      logger: deps.logger,
-      orchestrator: deps.orchestrator,
-      persistence: deps.persistence,
-      taskExecutor,
-      dispatchMode: deps.mutationTiming ? 'fire-and-forget' : 'await',
-      autoApproveAIFixes: deps.invokerConfig?.autoApproveAIFixes,
-      killRunningTask: async (taskId: string) => {
-        await taskExecutor.killActiveExecution(taskId);
-      },
-      commandService: deps.commandService,
-    }),
-    deleteWorkflow: async (workflowId: string) => {
-      const allTasks = deps.orchestrator.getAllTasks();
-      const workflowTasks = allTasks.filter(
-        (t) =>
-          t.config.workflowId === workflowId &&
-          (t.status === 'running' || t.status === 'fixing_with_ai'),
-      );
-      for (const task of workflowTasks) {
-        await taskExecutor.killActiveExecution(task.id);
-      }
-      await taskExecutor.closeWorkflowReview(workflowId);
-      const envelope = makeEnvelope('delete-workflow', 'headless', 'workflow', { workflowId });
-      const cmdResult = await deps.commandService.deleteWorkflow(envelope);
-      if (!cmdResult.ok) throw new Error(cmdResult.error.message);
-    },
-    detachWorkflow: async (workflowId: string, upstreamWorkflowId: string) => {
-      const envelope = makeEnvelope('detach-workflow', 'headless', 'workflow', { workflowId, upstreamWorkflowId });
-      const cmdResult = await deps.commandService.detachWorkflow(envelope);
-      if (!cmdResult.ok) throw new Error(cmdResult.error.message);
-    },
-  };
-}
-
-function buildHeadlessApproveAction(
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'commandService'>,
-  taskExecutor: TaskRunner,
-): (taskId: string) => Promise<{ started: TaskState[] }> {
-  return async (taskId: string) => {
-    const result = await approveTask(taskId, {
-      orchestrator: deps.orchestrator,
-      taskExecutor,
-      approve: async (approvedTaskId) => {
-        const envelope = makeEnvelope('approve', 'headless', 'task', { taskId: approvedTaskId });
-        const result = await deps.commandService.approve(envelope);
-        if (!result.ok) throw new Error(result.error.message);
-        return result.data;
-      },
-      resumeAfterFixApproval: async (approvedTaskId) => {
-        const envelope = makeEnvelope('approve', 'headless', 'task', { taskId: approvedTaskId });
-        const result = await deps.commandService.resumeTaskAfterFixApproval(envelope);
-        if (!result.ok) throw new Error(result.error.message);
-        return result.data;
-      },
-    });
-    return { started: result.started };
-  };
-}
-
-export function createHeadlessExecutor(
-  deps: HeadlessDeps,
-  callbackOverrides?: Partial<ConstructorParameters<typeof TaskRunner>[0]['callbacks']>,
-): TaskRunner {
-  const owner = deps.ownerTaskRunnerProvider?.() ?? null;
-  if (owner) {
-    if (callbackOverrides) {
-      deps.logger?.debug?.(
-        '[headless] createHeadlessExecutor: ignoring callbackOverrides — reusing owner TaskRunner',
-        { module: 'headless' },
-      );
-    }
-    return owner;
-  }
-  let executor: TaskRunner;
-  executor = new TaskRunner({
-    orchestrator: deps.orchestrator,
-    persistence: deps.persistence,
-    executorRegistry: deps.executorRegistry,
-    cwd: deps.repoRoot,
-    defaultBranch: deps.invokerConfig.defaultBranch,
-    dockerConfig: {
-      imageName: deps.invokerConfig.docker?.imageName,
-      secretsFile: resolveSecretsFilePath(deps.invokerConfig),
-    },
-    remoteTargetsProvider: () => loadConfig().remoteTargets ?? {},
-    executionPoolsProvider: () => deps.invokerConfig.executionPools ?? {},
-    onReviewGateCiFailure: deps.invokerConfig.autoFixCi
-      ? async (trigger) => {
-          await autoFixOnReviewGateFailure(trigger, {
-            orchestrator: deps.orchestrator,
-            persistence: deps.persistence,
-            taskExecutor: executor,
-            getAutoFixAgent: () => loadConfig().autoFixAgent,
-            getAutoApproveAIFixes: () => loadConfig().autoApproveAIFixes,
-          });
-        }
-      : undefined,
-    mergeGateProvider: new GitHubMergeGateProvider(),
-    reviewProviderRegistry: (() => {
-      const registry = new ReviewProviderRegistry();
-      registry.register(new GitHubMergeGateProvider());
-      return registry;
-    })(),
-    executionAgentRegistry: deps.executionAgentRegistry,
-    callbacks: {
-      onOutput: (taskId, data) => {
-        process.stdout.write(`\x1b[2m[${taskId}]\x1b[0m ${data}`);
-        try {
-          deps.persistence.appendTaskOutput(taskId, data);
-        } catch (err) {
-          deps.logger.error(`Failed to persist output for ${taskId}: ${err}`, { module: 'output' });
-        }
-      },
-      onHeartbeat: (taskId, event) => headlessHeartbeat(taskId, event, deps),
-      ...callbackOverrides,
-    },
-  });
-  return executor;
-}
+// ── Runnable-task dispatch (set/edit family) ─────────────────
 
 async function dispatchHeadlessRunnableTasks(
   deps: HeadlessDeps,
@@ -310,647 +159,6 @@ async function dispatchHeadlessRunnableTasks(
   const timer = setInterval(poll, 250);
   timer.unref?.();
   await new Promise<void>((resolve) => setImmediate(resolve));
-}
-
-export function wireHeadlessAutoFix(
-  deps: Pick<HeadlessDeps, 'logger' | 'messageBus' | 'orchestrator' | 'persistence' | 'commandService' | 'mutationTiming'>,
-  taskExecutor: Pick<TaskRunner, 'executeTasks' | 'fixWithAgent' | 'resolveConflict'>,
-  invokeAutoFix: (taskId: string) => Promise<void> = async (taskId) => {
-    const { autoFixOnFailure } = await import('./workflow-actions.js');
-    await autoFixOnFailure(taskId, {
-      logger: deps.logger,
-      orchestrator: deps.orchestrator,
-      persistence: deps.persistence,
-      commandService: deps.commandService,
-      taskExecutor: taskExecutor as TaskRunner,
-      mutationTiming: deps.mutationTiming,
-      getAutoFixAgent: () => loadConfig().autoFixAgent,
-      getAutoApproveAIFixes: () => loadConfig().autoApproveAIFixes,
-    });
-  },
-  onError: (taskId: string, err: unknown) => void = (taskId, err) => {
-    process.stderr.write(`[auto-fix] "${taskId}": ${err}\n`);
-  },
-): HeadlessAutoFixController {
-  const autoFixInProgress = new Set<string>();
-  const logHeadlessAutoFixDebug = (
-    taskId: string,
-    phase: string,
-    details: Record<string, unknown> = {},
-  ): void => {
-    const getTask = (deps.orchestrator as { getTask?: (id: string) => unknown }).getTask;
-    const task = getTask?.(taskId) as
-      | { status?: string; execution?: { autoFixAttempts?: number | null } }
-      | undefined;
-    const payload = {
-      phase,
-      status: task?.status ?? 'missing',
-      autoFixAttempts: task?.execution?.autoFixAttempts ?? null,
-      inProgressCount: autoFixInProgress.size,
-      inProgressForTask: autoFixInProgress.has(taskId),
-      ...details,
-    };
-    deps.persistence.logEvent?.(taskId, 'debug.auto-fix', payload);
-    process.stderr.write(`[auto-fix-debug][headless] task="${taskId}" phase=${phase} payload=${JSON.stringify(payload)}\n`);
-  };
-
-  const unsubscribe = deps.messageBus.subscribe<TaskDelta>(Channels.TASK_DELTA, (delta) => {
-    if (delta.type !== 'updated' || delta.changes.status !== 'failed') return;
-    const inProgress = autoFixInProgress.has(delta.taskId);
-    const shouldAutoFix = deps.orchestrator.shouldAutoFix(delta.taskId);
-    logHeadlessAutoFixDebug(delta.taskId, 'delta-failed', { shouldAutoFix, inProgress });
-    if (inProgress || !shouldAutoFix) {
-      logHeadlessAutoFixDebug(delta.taskId, 'schedule-skip', {
-        reason: !shouldAutoFix ? 'shouldAutoFix-false' : 'already-in-progress',
-      });
-      return;
-    }
-    autoFixInProgress.add(delta.taskId);
-    logHeadlessAutoFixDebug(delta.taskId, 'dispatch');
-    void invokeAutoFix(delta.taskId)
-      .catch((err) => {
-        logHeadlessAutoFixDebug(delta.taskId, 'dispatch-error', {
-          error: err instanceof Error ? err.stack ?? err.message : String(err),
-        });
-        onError(delta.taskId, err);
-      })
-      .finally(() => {
-        autoFixInProgress.delete(delta.taskId);
-        logHeadlessAutoFixDebug(delta.taskId, 'dispatch-finished');
-      });
-  });
-  return {
-    unsubscribe,
-    isBusy: () => autoFixInProgress.size > 0,
-  };
-}
-
-export function wireHeadlessApproveHook(deps: HeadlessDeps, te: TaskRunner): void {
-  deps.orchestrator.setBeforeApproveHook(async (task) => {
-    if (task.config.isMergeNode && task.config.workflowId && task.execution.pendingFixError === undefined) {
-      const workflow = deps.persistence.loadWorkflow(task.config.workflowId);
-      if (workflow?.mergeMode === "external_review") return;
-      await te.approveMerge(task.config.workflowId);
-    }
-  });
-}
-
-// ── Deprecation Warning ─────────────────────────────────────
-
-function warnDeprecated(oldCmd: string, newCmd: string): void {
-  process.stderr.write(
-    `${YELLOW}[deprecated]${RESET} "${oldCmd}" is deprecated. Use "${newCmd}" instead.\n`,
-  );
-}
-
-function assertDeleteAllEnabled(): void {
-  if (process.env.INVOKER_ALLOW_DELETE_ALL === '1') return;
-  throw new Error(
-    'delete-all is disabled by default. Set INVOKER_ALLOW_DELETE_ALL=1 to enable it explicitly.',
-  );
-}
-
-// ── Query Flag Parsing ──────────────────────────────────────
-
-export interface QueryFlags {
-  output: 'text' | 'label' | 'json' | 'jsonl';
-  status?: string;
-  workflow?: string;
-  noMerge?: boolean;
-  reset?: boolean;
-  groupBy?: string;
-  positional: string[];
-}
-
-export interface HeadlessAutoFixController {
-  unsubscribe: () => void;
-  isBusy: () => boolean;
-}
-
-export function parseQueryFlags(args: string[]): QueryFlags {
-  const flags: QueryFlags = { output: 'text', positional: [] };
-  let i = 0;
-  while (i < args.length) {
-    const arg = args[i];
-    if (arg === '--output' && i + 1 < args.length) {
-      const val = args[i + 1] as QueryFlags['output'];
-      if (!['text', 'label', 'json', 'jsonl'].includes(val)) {
-        throw new Error(`Invalid --output format: "${val}". Must be text|label|json|jsonl.`);
-      }
-      flags.output = val;
-      i += 2;
-    } else if (arg === '--status' && i + 1 < args.length) {
-      flags.status = args[i + 1];
-      i += 2;
-    } else if (arg === '--workflow' && i + 1 < args.length) {
-      flags.workflow = args[i + 1];
-      i += 2;
-    } else if (arg === '--no-merge') {
-      flags.noMerge = true;
-      i += 1;
-    } else if (arg === '--reset') {
-      flags.reset = true;
-      i += 1;
-    } else if (arg === '--group-by' && i + 1 < args.length) {
-      flags.groupBy = args[i + 1];
-      i += 2;
-    } else if (arg.startsWith('--')) {
-      throw new Error(`Unknown query flag: "${arg}"`);
-    } else {
-      flags.positional.push(arg);
-      i += 1;
-    }
-  }
-  return flags;
-}
-
-// ── Query Router ────────────────────────────────────────────
-
-async function headlessQuery(args: string[], deps: HeadlessDeps): Promise<void> {
-  const subCommand = args[0];
-  if (!subCommand) {
-    throw new Error('Missing query sub-command. Usage: --headless query <workflows|tasks|task|queue|audit|session|cost|cost-events|costs|ui-perf|stats>');
-  }
-  const flags = parseQueryFlags(args.slice(1));
-
-  const {
-    formatWorkflowList, formatTaskStatus, formatWorkflowStatus,
-    formatEventLog, formatQueueStatus, formatWorkflowStats,
-    serializeWorkflow, serializeTask, serializeEvent,
-    formatAsLabel, formatAsJson, formatAsJsonl,
-  } = await import('./formatter.js');
-
-  switch (subCommand) {
-    case 'workflows': {
-      let workflows = deps.persistence.listWorkflows();
-      if (flags.status) {
-        workflows = workflows.filter(wf => wf.status === flags.status);
-      }
-      switch (flags.output) {
-        case 'label': process.stdout.write(formatAsLabel(workflows) + '\n'); break;
-        case 'json':  process.stdout.write(formatAsJson(workflows.map(serializeWorkflow)) + '\n'); break;
-        case 'jsonl': process.stdout.write(formatAsJsonl(workflows.map(serializeWorkflow)) + '\n'); break;
-        default:      process.stdout.write(formatWorkflowList(workflows) + '\n'); break;
-      }
-      break;
-    }
-    case 'workflow': {
-      const workflowId = flags.positional[0];
-      if (!workflowId) throw new Error('Missing workflowId. Usage: --headless query workflow <workflowId>');
-      const workflow = deps.persistence.loadWorkflow(workflowId);
-      if (!workflow) throw new Error(`Workflow "${workflowId}" not found.`);
-      switch (flags.output) {
-        case 'label': process.stdout.write(`${workflow.id}\n`); break;
-        case 'json':  process.stdout.write(formatAsJson(serializeWorkflow(workflow)) + '\n'); break;
-        case 'jsonl': process.stdout.write(formatAsJsonl([serializeWorkflow(workflow)]) + '\n'); break;
-        default:      process.stdout.write(formatWorkflowList([workflow]) + '\n'); break;
-      }
-      break;
-    }
-    case 'tasks': {
-      const { orchestrator, persistence } = deps;
-      const workflows = persistence.listWorkflows();
-      if (workflows.length === 0) {
-        process.stdout.write('No workflows found. Run a plan first.\n');
-        return;
-      }
-
-      // Support both:
-      //   query tasks --workflow <id>
-      //   query tasks <workflowId>
-      const workflowFilter = flags.workflow ?? flags.positional[0];
-
-      // Load tasks from specific workflow or latest
-      const targetWorkflows = workflowFilter
-        ? workflows.filter(wf => wf.id === workflowFilter)
-        : [workflows[0]];
-
-      if (targetWorkflows.length === 0) {
-        throw new Error(`Workflow "${workflowFilter}" not found.`);
-      }
-
-      let allTasks: import('@invoker/workflow-core').TaskState[] = [];
-      for (const wf of targetWorkflows) {
-        // Query must stay read-only: sync graph from DB without starting/restarting tasks.
-        orchestrator.syncFromDb(wf.id);
-        // Filter by workflow ID — the orchestrator may have loaded other workflows during init.
-        allTasks.push(...orchestrator.getAllTasks().filter(t => t.config.workflowId === wf.id));
-      }
-
-      // Apply filters
-      if (flags.status) {
-        allTasks = allTasks.filter(t => t.status === flags.status);
-      }
-      if (flags.noMerge) {
-        allTasks = allTasks.filter(t => !t.config.isMergeNode);
-      }
-
-      switch (flags.output) {
-        case 'label': process.stdout.write(formatAsLabel(allTasks) + '\n'); break;
-        case 'json':  process.stdout.write(formatAsJson(allTasks.map(serializeTask)) + '\n'); break;
-        case 'jsonl': process.stdout.write(formatAsJsonl(allTasks.map(serializeTask)) + '\n'); break;
-        default: {
-          for (const task of allTasks) process.stdout.write(formatTaskStatus(task) + '\n');
-          const status = orchestrator.getWorkflowStatus();
-          process.stdout.write(`\n${formatWorkflowStatus(status)}\n`);
-          break;
-        }
-      }
-      break;
-    }
-    case 'task': {
-      const taskId = flags.positional[0];
-      if (!taskId) throw new Error('Usage: --headless query task <taskId>');
-      const resolved = restoreWorkflowForTask(taskId, deps).resolvedTaskId;
-      const task = deps.orchestrator.getTask(resolved);
-      if (!task) throw new Error(`Task "${taskId}" not found`);
-
-      switch (flags.output) {
-        case 'label': process.stdout.write(task.id + '\n'); break;
-        case 'json':  process.stdout.write(formatAsJson(serializeTask(task)) + '\n'); break;
-        case 'jsonl': process.stdout.write(formatAsJsonl([serializeTask(task)]) + '\n'); break;
-        default:      process.stdout.write(task.status + '\n'); break;
-      }
-      break;
-    }
-    case 'queue': {
-      const workflows = deps.persistence.listWorkflows();
-      for (const workflow of workflows) {
-        deps.orchestrator.syncFromDb(workflow.id);
-      }
-      const status = deps.orchestrator.getQueueStatus();
-
-      switch (flags.output) {
-        case 'label': {
-          const ids = [...status.running.map(t => t.taskId), ...status.queued.map(t => t.taskId)];
-          process.stdout.write(ids.join('\n') + '\n');
-          break;
-        }
-        case 'json':  process.stdout.write(formatAsJson(status) + '\n'); break;
-        case 'jsonl': {
-          for (const t of status.running) process.stdout.write(JSON.stringify({ ...t, state: 'running' }) + '\n');
-          for (const t of status.queued) process.stdout.write(JSON.stringify({ ...t, state: 'queued' }) + '\n');
-          break;
-        }
-        default: process.stdout.write(formatQueueStatus(status) + '\n'); break;
-      }
-      break;
-    }
-    case 'audit': {
-      const taskId = flags.positional[0];
-      if (!taskId) throw new Error('Usage: --headless query audit <taskId>');
-      const events = deps.persistence.getEvents(taskId);
-
-      switch (flags.output) {
-        case 'label': process.stdout.write(events.map(e => `${e.taskId}:${e.eventType}`).join('\n') + '\n'); break;
-        case 'json':  process.stdout.write(formatAsJson(events.map(serializeEvent)) + '\n'); break;
-        case 'jsonl': process.stdout.write(formatAsJsonl(events.map(serializeEvent)) + '\n'); break;
-        default:      process.stdout.write(formatEventLog(events) + '\n'); break;
-      }
-      break;
-    }
-    case 'session': {
-      const taskId = flags.positional[0];
-      if (!taskId) throw new Error('Usage: --headless query session <taskId>');
-      // For non-text output, we'd need structured session data.
-      // For now, session only supports text output; other formats fall through to text.
-      await headlessSession(taskId, deps);
-      break;
-    }
-    case 'ui-perf': {
-      if (flags.reset) {
-        deps.resetUiPerfStats?.();
-      }
-      const stats = deps.getUiPerfStats?.() ?? {
-        ownerMode: 'local',
-        ts: new Date().toISOString(),
-        mainDeltaToUi: 0,
-        dbPollCreated: 0,
-        dbPollUpdatedAsCreated: 0,
-        dbPollUpdatedAsUpdated: 0,
-        rendererReports: 0,
-        maxRendererEventLoopLagMs: 0,
-        maxRendererLongTaskMs: 0,
-      };
-      switch (flags.output) {
-        case 'label':
-          process.stdout.write(String((stats as Record<string, unknown>).maxRendererEventLoopLagMs ?? 0) + '\n');
-          break;
-        case 'json':
-          process.stdout.write(formatAsJson(stats) + '\n');
-          break;
-        case 'jsonl':
-          process.stdout.write(formatAsJsonl([stats]) + '\n');
-          break;
-        default:
-          process.stdout.write(`${JSON.stringify(stats, null, 2)}\n`);
-          break;
-      }
-      break;
-    }
-    case 'stats': {
-      const { orchestrator, persistence } = deps;
-      const workflows = persistence.listWorkflows();
-
-      const completed = workflows.filter(w => w.status === 'completed').length;
-      const failed = workflows.filter(w => w.status === 'failed').length;
-      const running = workflows.filter(w => w.status === 'running').length;
-      const terminal = completed + failed;
-      const successRate = terminal > 0 ? (completed / terminal) * 100 : 0;
-
-      // Average duration across workflows that have both timestamps.
-      // startedAt/completedAt are added by the workflow-duration feature; guard for older DBs.
-      const durations = (workflows as Array<typeof workflows[0] & { startedAt?: string; completedAt?: string }>)
-        .filter(w => w.startedAt && w.completedAt)
-        .map(w => new Date(w.completedAt!).getTime() - new Date(w.startedAt!).getTime());
-      const avgDurationMs = durations.length > 0
-        ? durations.reduce((a, b) => a + b, 0) / durations.length
-        : null;
-
-      // Most-failed task descriptions across all workflows
-      const failCounts = new Map<string, number>();
-      for (const wf of workflows) {
-        orchestrator.syncFromDb(wf.id);
-        const tasks = orchestrator.getAllTasks().filter(
-          t => t.config.workflowId === wf.id && !t.config.isMergeNode && t.status === 'failed',
-        );
-        for (const t of tasks) {
-          failCounts.set(t.description, (failCounts.get(t.description) ?? 0) + 1);
-        }
-      }
-      const mostFailedTasks = [...failCounts.entries()]
-        .sort((a, b) => b[1] - a[1])
-        .slice(0, 5)
-        .map(([description, failCount]) => ({ description, failCount }));
-
-      const stats = {
-        totalWorkflows: workflows.length,
-        completed,
-        failed,
-        running,
-        successRate,
-        avgDurationMs,
-        mostFailedTasks,
-      };
-
-      switch (flags.output) {
-        case 'label': process.stdout.write(`${successRate.toFixed(1)}%\n`); break;
-        case 'json':  process.stdout.write(formatAsJson(stats) + '\n'); break;
-        case 'jsonl': process.stdout.write(formatAsJsonl([stats]) + '\n'); break;
-        default:      process.stdout.write(formatWorkflowStats(stats) + '\n'); break;
-      }
-      break;
-    }
-    case 'cost': {
-      await headlessCost(flags, deps);
-      break;
-    }
-    case 'cost-events': {
-      await headlessCostEvents(flags, deps);
-      break;
-    }
-    case 'costs': {
-      await headlessCosts(flags, deps);
-      break;
-    }
-    default:
-      throw new Error(`Unknown query sub-command: "${subCommand}". Use: workflows, tasks, task, queue, audit, session, cost, cost-events, costs, ui-perf, stats`);
-  }
-}
-
-// ── Cost Query ──────────────────────────────────────────────
-
-async function headlessCosts(
-  flags: QueryFlags,
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'persistence' | 'executionAgentRegistry'>,
-): Promise<void> {
-  const {
-    formatGroupedCostRollups, formatCostRollup,
-    serializeCostEvent,
-    formatAsJson,
-  } = await import('./formatter.js');
-  const {
-    groupCostEvents,
-    serializeGroupedRollup,
-  } = await import('./cost-rollup.js');
-  const { rollUpCostEvents } = await import('@invoker/contracts');
-  const allEvents = await collectCostEvents(flags, deps);
-
-  if (allEvents.length === 0) {
-    process.stdout.write('No cost data available.\n');
-    return;
-  }
-
-  // Group by all dimensions and output
-  const grouped = groupCostEvents(allEvents);
-  const totalRollup = rollUpCostEvents(allEvents);
-
-  switch (flags.output) {
-    case 'label':
-      process.stdout.write(`${totalRollup.totalTokens} tokens $${totalRollup.totalCostUsd.toFixed(4)}\n`);
-      break;
-    case 'json':
-      process.stdout.write(formatAsJson({
-        groups: grouped.map(serializeGroupedRollup),
-        total: totalRollup,
-        events: allEvents.map(serializeCostEvent),
-      }) + '\n');
-      break;
-    case 'jsonl':
-      for (const event of allEvents) {
-        process.stdout.write(JSON.stringify(serializeCostEvent(event)) + '\n');
-      }
-      break;
-    default:
-      process.stdout.write(formatGroupedCostRollups(grouped) + '\n');
-      process.stdout.write('\n');
-      process.stdout.write(formatCostRollup(totalRollup) + '\n');
-      break;
-  }
-}
-
-// ── Shared Cost Event Collection ────────────────────────────
-
-type CostQueryDeps = Pick<HeadlessDeps, 'orchestrator' | 'persistence' | 'executionAgentRegistry'>;
-
-function resolveCostAttributionAttempt(
-  task: TaskState,
-  attempts: readonly Attempt[],
-): Attempt | undefined {
-  const sessionId = task.execution.agentSessionId?.trim();
-  if (sessionId) {
-    const exactSessionAttempt = attempts.find((attempt) => attempt.agentSessionId?.trim() === sessionId);
-    if (exactSessionAttempt) return exactSessionAttempt;
-  }
-
-  const selectedAttemptId = task.execution.selectedAttemptId?.trim();
-  if (selectedAttemptId) {
-    const selectedAttempt = attempts.find((attempt) => attempt.id === selectedAttemptId);
-    if (selectedAttempt) return selectedAttempt;
-  }
-
-  return attempts.at(-1);
-}
-
-async function collectCostEvents(
-  flags: QueryFlags,
-  deps: CostQueryDeps,
-): Promise<import('@invoker/contracts').NormalizedCostEvent[]> {
-  const { attributeSessionUsage, buildAttributionContext } = await import('./cost-rollup.js');
-
-  const workflowFilter = flags.workflow ?? flags.positional[0];
-  const workflows = deps.persistence.listWorkflows();
-  if (workflows.length === 0) return [];
-
-  const targetWorkflows = workflowFilter
-    ? workflows.filter(wf => wf.id === workflowFilter)
-    : workflows;
-
-  if (workflowFilter && targetWorkflows.length === 0) {
-    throw new Error(`Workflow "${workflowFilter}" not found.`);
-  }
-
-  const allEvents: import('@invoker/contracts').NormalizedCostEvent[] = [];
-
-  for (const wf of targetWorkflows) {
-    deps.orchestrator.syncFromDb(wf.id);
-    const tasks = deps.orchestrator.getAllTasks().filter(
-      t => t.config.workflowId === wf.id && !t.config.isMergeNode,
-    );
-
-    for (const task of tasks) {
-      const attempts = deps.persistence.loadAttempts(task.id);
-      const attributedAttempt = resolveCostAttributionAttempt(task, attempts);
-      const ctx = buildAttributionContext({
-        id: task.id,
-        workflowId: wf.id,
-        runnerKind: task.config.runnerKind ?? 'worktree',
-        agentSessionId: task.execution.agentSessionId,
-        lastAgentSessionId: task.execution.lastAgentSessionId,
-        agentName: task.execution.agentName,
-        lastAgentName: task.execution.lastAgentName,
-      }, attributedAttempt?.id ?? task.execution.selectedAttemptId?.trim() ?? task.id, attributedAttempt?.agentSessionId?.trim());
-      if (!ctx) continue;
-
-      const agentName = ctx.agentName;
-      const driver = deps.executionAgentRegistry?.getSessionDriver(agentName);
-      if (!driver?.extractUsage) continue;
-
-      const raw = driver.loadSession(ctx.agentSessionId);
-      if (!raw) continue;
-
-      const usageEvents = driver.extractUsage(raw);
-      const attributed = attributeSessionUsage(usageEvents, ctx);
-      allEvents.push(...attributed);
-    }
-  }
-
-  return allEvents;
-}
-
-// ── query cost ──────────────────────────────────────────────
-
-const VALID_GROUP_DIMENSIONS = ['workflow', 'task', 'agent', 'model', 'day'] as const;
-
-async function headlessCost(
-  flags: QueryFlags,
-  deps: CostQueryDeps,
-): Promise<void> {
-  const {
-    formatGroupedCostRollups, formatCostRollup,
-    formatAsJson,
-  } = await import('./formatter.js');
-  const { groupCostEvents, serializeGroupedRollup } = await import('./cost-rollup.js');
-  const { rollUpCostEvents } = await import('@invoker/contracts');
-
-  // Parse --group-by flag (comma-separated dimensions)
-  let dimensions: CostGroupDimension[] | undefined;
-  if (flags.groupBy) {
-    const parts = flags.groupBy.split(',').map(s => s.trim());
-    for (const part of parts) {
-      if (!(VALID_GROUP_DIMENSIONS as readonly string[]).includes(part)) {
-        throw new Error(
-          `Invalid --group-by dimension: "${part}". Must be one or more of: ${VALID_GROUP_DIMENSIONS.join(', ')}`,
-        );
-      }
-    }
-    dimensions = parts as CostGroupDimension[];
-  }
-
-  const allEvents = await collectCostEvents(flags, deps);
-
-  if (allEvents.length === 0) {
-    process.stdout.write('No cost data available.\n');
-    return;
-  }
-
-  const grouped = groupCostEvents(allEvents, dimensions);
-  const totals = rollUpCostEvents(allEvents);
-  const scope = flags.workflow ?? flags.positional[0] ?? 'all';
-  const groupBy = dimensions ?? [...VALID_GROUP_DIMENSIONS];
-
-  switch (flags.output) {
-    case 'label':
-      process.stdout.write(`${totals.totalTokens} tokens $${totals.totalCostUsd.toFixed(4)}\n`);
-      break;
-    case 'json':
-      process.stdout.write(formatAsJson({
-        scope,
-        groupBy,
-        totals,
-        groups: grouped.map(serializeGroupedRollup),
-        metadata: { eventCount: allEvents.length },
-      }) + '\n');
-      break;
-    case 'jsonl':
-      for (const group of grouped) {
-        process.stdout.write(JSON.stringify(serializeGroupedRollup(group)) + '\n');
-      }
-      break;
-    default:
-      process.stdout.write(formatGroupedCostRollups(grouped) + '\n');
-      process.stdout.write('\n');
-      process.stdout.write(formatCostRollup(totals) + '\n');
-      break;
-  }
-}
-
-// ── query cost-events ───────────────────────────────────────
-
-async function headlessCostEvents(
-  flags: QueryFlags,
-  deps: CostQueryDeps,
-): Promise<void> {
-  const {
-    formatCostEvent, serializeCostEvent,
-    formatAsJson, formatAsJsonl,
-  } = await import('./formatter.js');
-
-  const allEvents = await collectCostEvents(flags, deps);
-
-  if (allEvents.length === 0) {
-    process.stdout.write('No cost events found.\n');
-    return;
-  }
-
-  switch (flags.output) {
-    case 'label':
-      for (const event of allEvents) {
-        process.stdout.write(`${event.attribution.taskId}:${event.identity.eventId}\n`);
-      }
-      break;
-    case 'json':
-      process.stdout.write(formatAsJson(allEvents.map(serializeCostEvent)) + '\n');
-      break;
-    case 'jsonl':
-      process.stdout.write(formatAsJsonl(allEvents.map(serializeCostEvent)) + '\n');
-      break;
-    default:
-      for (const event of allEvents) {
-        process.stdout.write(formatCostEvent(event) + '\n');
-      }
-      break;
-  }
 }
 
 // ── Set Router ──────────────────────────────────────────────
@@ -1353,988 +561,7 @@ ${BOLD}Options:${RESET}
 `);
 }
 
-async function trackHeadlessWorkflow(
-  workflowId: string,
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'messageBus'>,
-  options: {
-    waitForApproval?: boolean;
-    hasBackgroundWork?: () => boolean;
-    printSnapshot?: boolean;
-    printSummary?: boolean;
-    printTaskOutput?: boolean;
-    allowSignals?: boolean;
-    syncFromDb?: boolean;
-    setExitCodeOnFailure?: boolean;
-  } = {},
-): Promise<Awaited<ReturnType<typeof trackWorkflow>>> {
-  if (options.waitForApproval) {
-    process.stdout.write('[headless] Waiting for PR approval (--wait-for-approval)...\n');
-  }
-  return await trackWorkflow({
-    workflowId,
-    messageBus: deps.messageBus,
-    waitForApproval: options.waitForApproval,
-    hasBackgroundWork: options.hasBackgroundWork,
-    printSnapshot: options.printSnapshot,
-    printSummary: options.printSummary,
-    printTaskOutput: options.printTaskOutput,
-    allowSignals: options.allowSignals,
-    setExitCodeOnFailure: options.setExitCodeOnFailure,
-    maxWaitMs: options.allowSignals ? undefined : (options.waitForApproval ? 86_400_000 : 1_800_000),
-    loadTasks: () => {
-      if (options.syncFromDb) {
-        deps.orchestrator.syncFromDb(workflowId);
-      }
-      return deps.orchestrator.getAllTasks().filter((task) => task.config.workflowId === workflowId);
-    },
-  });
-}
-
-async function headlessWatch(workflowId: string | undefined, deps: HeadlessDeps): Promise<void> {
-  const workflows = deps.persistence.listWorkflows();
-  if (workflows.length === 0) {
-    process.stdout.write('No workflows found. Run a plan first.\n');
-    return;
-  }
-  const targetWorkflowId = workflowId ?? workflows[0]?.id;
-  const workflow = workflows.find((item) => item.id === targetWorkflowId);
-  if (!workflow || !targetWorkflowId) {
-    throw new Error(`Workflow "${workflowId}" not found.`);
-  }
-
-  process.stdout.write(`${BOLD}Watching workflow: ${workflow.id}${RESET}\n\n`);
-  const result = await trackHeadlessWorkflow(workflow.id, deps, {
-    printSnapshot: true,
-    printSummary: true,
-    printTaskOutput: false,
-    allowSignals: true,
-    syncFromDb: true,
-    setExitCodeOnFailure: true,
-  });
-  process.stdout.write(`\n[watch] done — ${result.status.completed} completed, ${result.status.failed} failed, ${result.status.closed} closed\n`);
-}
-
-// ── Headless Commands ────────────────────────────────────────
-
-async function headlessRun(
-  planPath: string,
-  deps: HeadlessDeps,
-  waitForApproval?: boolean,
-  noTrack?: boolean,
-): Promise<void> {
-  const { orchestrator, repoRoot, invokerConfig } = deps;
-  if (!planPath) throw new Error('Missing plan file. Usage: --headless run <plan.yaml>');
-
-  const { readFile } = await import('node:fs/promises');
-  const { parsePlanFile } = await import('./plan-parser.js');
-
-  const yamlSource = await readFile(planPath, 'utf-8');
-  const plan = await parsePlanFile(planPath);
-  const execRegistry = deps.executionAgentRegistry ?? registerBuiltinAgents();
-  assertPlanExecutionAgentsRegistered(plan, execRegistry);
-  backupPlan(plan, yamlSource, deps.logger);
-  process.stdout.write(`${BOLD}Loading plan: ${plan.name}${RESET}\n`);
-  process.stdout.write(`Tasks: ${plan.tasks.length}\n\n`);
-
-  const taskExecutor = createHeadlessExecutor(deps);
-  wireHeadlessApproveHook(deps, taskExecutor);
-  const autoFix = wireHeadlessAutoFix(deps, taskExecutor);
-
-  const api = startApiServer({
-    logger: deps.logger,
-    orchestrator,
-    persistence: deps.persistence,
-    executorRegistry: deps.executorRegistry,
-    ...buildHeadlessApiServerDeps(deps, taskExecutor),
-  });
-
-  const wfIdsBefore = new Set(orchestrator.getWorkflowIds());
-  orchestrator.loadPlan(plan, { allowGraphMutation: invokerConfig.allowGraphMutation });
-  const currentWorkflowId = orchestrator.getWorkflowIds().find((id) => !wfIdsBefore.has(id));
-  if (currentWorkflowId) process.stdout.write(`Workflow ID: ${currentWorkflowId}\n`);
-
-  const started = orchestrator.startExecution();
-
-  if (noTrack) {
-    if (started.length > 0) {
-      void Promise.resolve()
-        .then(() => taskExecutor.executeTasks(started))
-        .catch((err) => {
-          deps.logger.error(
-            `background no-track run failed for ${currentWorkflowId ?? 'unknown'}: ${err instanceof Error ? err.stack ?? err.message : String(err)}`,
-            { module: 'headless' },
-          );
-        });
-    }
-    process.stdout.write('[headless] --no-track enabled: submission accepted; exiting without tracking.\n');
-    await api.close().catch(() => {});
-    return;
-  }
-
-  if (started.length > 0) {
-    await taskExecutor.executeTasks(started);
-  }
-
-  if (currentWorkflowId) {
-    await trackHeadlessWorkflow(currentWorkflowId, deps, {
-      waitForApproval,
-      hasBackgroundWork: autoFix.isBusy,
-      printSnapshot: true,
-      printSummary: true,
-      printTaskOutput: true,
-      setExitCodeOnFailure: true,
-    });
-  }
-
-  await api.close().catch(() => {});
-  autoFix.unsubscribe();
-}
-
-async function headlessResume(
-  workflowId: string,
-  deps: HeadlessDeps,
-  waitForApproval?: boolean,
-  noTrack?: boolean,
-): Promise<void> {
-  const { orchestrator } = deps;
-  if (!workflowId) throw new Error('Missing workflowId. Usage: --headless resume <id>');
-
-  process.stdout.write(`${BOLD}Resuming workflow: ${workflowId}${RESET}\n\n`);
-
-  const taskExecutor = createHeadlessExecutor(deps);
-  wireHeadlessApproveHook(deps, taskExecutor);
-  const autoFix = wireHeadlessAutoFix(deps, taskExecutor);
-
-  const api = startApiServer({
-    logger: deps.logger,
-    orchestrator,
-    persistence: deps.persistence,
-    executorRegistry: deps.executorRegistry,
-    ...buildHeadlessApiServerDeps(deps, taskExecutor),
-  });
-
-  orchestrator.syncFromDb(workflowId);
-  const allStarted = orchestrator.startExecution();
-
-  if (noTrack) {
-    if (allStarted.length > 0) {
-      void Promise.resolve()
-        .then(() => taskExecutor.executeTasks(allStarted))
-        .catch((err) => {
-          deps.logger.error(
-            `background no-track resume failed for ${workflowId}: ${err instanceof Error ? err.stack ?? err.message : String(err)}`,
-            { module: 'headless' },
-          );
-        });
-    }
-    process.stdout.write('[headless] --no-track enabled: resume accepted; exiting without tracking.\n');
-    await api.close().catch(() => {});
-    autoFix.unsubscribe();
-    return;
-  }
-
-  if (allStarted.length === 0) {
-    await api.close().catch(() => {});
-    autoFix.unsubscribe();
-    return;
-  }
-
-  await taskExecutor.executeTasks(allStarted);
-
-  await trackHeadlessWorkflow(workflowId, deps, {
-    waitForApproval,
-    hasBackgroundWork: autoFix.isBusy,
-    printSnapshot: true,
-    printSummary: true,
-    printTaskOutput: true,
-    setExitCodeOnFailure: true,
-  });
-
-  await api.close().catch(() => {});
-  autoFix.unsubscribe();
-}
-
-async function headlessApprove(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId) throw new Error('Missing taskId.');
-  await withRestoredTaskUnlessDeleteAllWon(taskId, deps, 'approve', async (restored) => {
-    taskId = restored.resolvedTaskId;
-    const te = createHeadlessExecutor(deps);
-    wireHeadlessApproveHook(deps, te);
-    const autoFix = wireHeadlessAutoFix(deps, te);
-    const approveTaskAction = buildHeadlessApproveAction(deps, te);
-    const beforeStatus = deps.orchestrator.getWorkflowStatus(restored.workflowId);
-    const { started } = await approveTaskAction(taskId);
-    await finalizeMutationWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.approve',
-      started,
-      mutationTiming: deps.mutationTiming,
-      scopedTaskIds: [taskId],
-    });
-    process.stdout.write(`Approved task: ${taskId}\n`);
-    if (deps.noTrack) {
-      process.stdout.write('[headless] --no-track enabled: approve accepted; exiting without tracking.\n');
-      autoFix.unsubscribe();
-      return;
-    }
-    const afterStatus = deps.orchestrator.getWorkflowStatus(restored.workflowId);
-    const workflowTasks = deps
-      .orchestrator
-      .getAllTasks()
-      .filter((task) => task.config.workflowId === restored.workflowId);
-    const readyTasks = (deps.orchestrator.getReadyTasks?.() ?? [])
-      .filter((task) => task.config.workflowId === restored.workflowId && task.status === 'pending');
-    if (readyTasks.length > 0) {
-      await executeGlobalTopup({
-        orchestrator: deps.orchestrator,
-        taskExecutor: te,
-        logger: deps.logger,
-        context: 'headless.approve.ready-tasks',
-        mutationTiming: deps.mutationTiming,
-      });
-    }
-    const hasRunningWork = workflowTasks.some(
-      (task) => task.status === 'running' || task.status === 'fixing_with_ai',
-    );
-    const resumedWork =
-      hasRunningWork
-      || afterStatus.running > beforeStatus.running
-      || afterStatus.pending < beforeStatus.pending
-      || readyTasks.length > 0;
-    if (!resumedWork) {
-      autoFix.unsubscribe();
-      return;
-    }
-    await trackHeadlessWorkflow(restored.workflowId, deps, {
-      hasBackgroundWork: autoFix.isBusy,
-      printSummary: false,
-      printTaskOutput: true,
-      setExitCodeOnFailure: false,
-    });
-    autoFix.unsubscribe();
-  });
-}
-
-async function headlessReject(taskId: string, deps: Pick<HeadlessDeps, 'commandService' | 'orchestrator' | 'persistence'>, reason?: string): Promise<void> {
-  if (!taskId) throw new Error('Missing taskId.');
-  await withRestoredTaskUnlessDeleteAllWon(taskId, deps, 'reject', async (restored) => {
-    taskId = restored.resolvedTaskId;
-    const envelope = makeEnvelope('reject', 'headless', 'task', { taskId, reason });
-    const result = await deps.commandService.reject(envelope);
-    if (!result.ok) throw new Error(result.error.message);
-    process.stdout.write(`Rejected task: ${taskId}${reason ? ` (reason: ${reason})` : ''}\n`);
-  });
-}
-
-async function headlessInput(taskId: string, text: string, deps: Pick<HeadlessDeps, 'commandService' | 'orchestrator' | 'persistence'>): Promise<void> {
-  if (!taskId || !text) throw new Error('Missing arguments. Usage: --headless input <taskId> <text>');
-  await withRestoredTaskUnlessDeleteAllWon(taskId, deps, 'input', async (restored) => {
-    taskId = restored.resolvedTaskId;
-    const envelope = makeEnvelope('provide-input', 'headless', 'task', { taskId, input: text });
-    const result = await deps.commandService.provideInput(envelope);
-    if (!result.ok) throw new Error(result.error.message);
-    process.stdout.write(`Input provided to task: ${taskId}\n`);
-  });
-}
-
-async function headlessSelect(taskId: string, experimentId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId || !experimentId) throw new Error('Missing arguments. Usage: --headless select <taskId> <expId>');
-  await withRestoredTaskUnlessDeleteAllWon(taskId, deps, 'select', async ({ workflowId, resolvedTaskId }) => {
-    const envelope = makeEnvelope('select-experiment', 'headless', 'task', { taskId: resolvedTaskId, experimentId });
-    const result = await deps.commandService.selectExperiment(envelope);
-    if (!result.ok) throw new Error(result.error.message);
-    process.stdout.write(`Selected experiment ${experimentId} for task: ${resolvedTaskId}\n`);
-
-    const taskExecutor = createHeadlessExecutor(deps);
-    const autoFix = wireHeadlessAutoFix(deps, taskExecutor);
-    const started = deps.orchestrator.resumeWorkflow(workflowId);
-    void started;
-    await trackHeadlessWorkflow(workflowId, deps, {
-      hasBackgroundWork: autoFix.isBusy,
-      printSummary: false,
-      printTaskOutput: true,
-      setExitCodeOnFailure: false,
-    });
-    autoFix.unsubscribe();
-  });
-}
-
-async function headlessRetryTask(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId) throw new Error('Missing arguments. Usage: --headless retry-task <taskId>');
-  await withRestoredTaskUnlessDeleteAllWon(taskId, deps, 'retry-task', async (restored) => {
-    taskId = restored.resolvedTaskId;
-    if (deps.mutationTiming) {
-      await deps.mutationTiming.span(
-        'headless.retry-task.preemptTaskSubgraph',
-        { taskId },
-        () => preemptTaskSubgraph(taskId, deps),
-      );
-    } else {
-      await preemptTaskSubgraph(taskId, deps);
-    }
-
-    const envelope = makeEnvelope('restart-task', 'headless', 'task', { taskId });
-    const result = deps.mutationTiming
-      ? await deps.mutationTiming.span(
-        'headless.retry-task.commandService.retryTask',
-        { taskId },
-        () => deps.commandService.retryTask(envelope),
-      )
-      : await deps.commandService.retryTask(envelope);
-    if (!result.ok) throw new Error(result.error.message);
-    const runnable = result.data.filter(isDispatchableLaunch);
-    process.stdout.write(`Restarted task "${taskId}" — ${runnable.length} task(s) to execute\n`);
-
-    if (deps.noTrack) {
-      const runningKey = (task: TaskState): string => {
-        const attemptId = task.execution.selectedAttemptId?.trim();
-        return attemptId ? `attempt:${attemptId}` : `task:${task.id}`;
-      };
-      const scopedKeys = new Set(runnable.map((task) => runningKey(task)));
-      const globalTopup = deps.orchestrator
-        .startExecution()
-        .filter(isDispatchableLaunch)
-        .filter((task) => !scopedKeys.has(runningKey(task)));
-      const dispatchable = [...runnable, ...globalTopup];
-      if (dispatchable.length > 0) {
-        if (deps.deferRunnableTasks) {
-          deps.deferRunnableTasks(dispatchable, restored.workflowId);
-        } else {
-          const taskExecutor = createHeadlessExecutor(deps);
-          const launch = setTimeout(() => {
-            void taskExecutor.executeTasks(dispatchable).catch((err) => {
-              deps.logger.error(
-                `background no-track task retry failed for ${taskId}: ${err instanceof Error ? err.stack ?? err.message : String(err)}`,
-                { module: 'headless' },
-              );
-            });
-          }, 25);
-          launch.unref?.();
-        }
-      }
-      process.stdout.write('[headless] --no-track enabled: retry-task accepted; exiting without tracking.\n');
-      return;
-    }
-
-    const taskExecutor = createHeadlessExecutor(deps);
-    const autoFix = wireHeadlessAutoFix(deps, taskExecutor);
-    const { topup } = await dispatchStartedTasksWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor,
-      logger: deps.logger,
-      context: 'headless.restart-task',
-      started: result.data,
-      scopedTaskIds: [taskId],
-      mutationTiming: deps.mutationTiming,
-    });
-    if (runnable.length + topup.length === 0) {
-      autoFix.unsubscribe();
-      return;
-    }
-    await trackHeadlessWorkflow(restored.workflowId, deps, {
-      hasBackgroundWork: autoFix.isBusy,
-      printSummary: false,
-      printTaskOutput: true,
-      setExitCodeOnFailure: false,
-    });
-    autoFix.unsubscribe();
-  });
-}
-
-async function headlessFix(rawArgs: string[], deps: HeadlessDeps): Promise<void> {
-  const parsed = parseHeadlessFixArgs(rawArgs);
-  let taskId = parsed.taskId;
-  if (!taskId) throw new Error('Missing taskId. Usage: --headless fix <taskId> [claude|codex] [--auto-fix]');
-  const restored = restoreWorkflowForTaskUnlessDeleteAllWon(taskId, deps, 'fix');
-  if (!restored) return;
-  taskId = restored.resolvedTaskId;
-
-  if (parsed.autoFix) {
-    const task = deps.orchestrator.getTask(taskId);
-    const attemptsBefore = task?.execution.autoFixAttempts ?? 0;
-    deps.persistence.updateTask(taskId, { execution: { autoFixAttempts: attemptsBefore + 1 } });
-  }
-
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  const agent = (parsed.agentName ?? 'claude').toLowerCase();
-  try {
-    const result = await fixWithAgentAction(taskId, {
-      logger: deps.logger,
-      orchestrator: deps.orchestrator,
-      persistence: deps.persistence,
-      commandService: deps.commandService,
-      taskExecutor: te,
-      mutationTiming: deps.mutationTiming,
-      autoApproveAIFixes: deps.invokerConfig.autoApproveAIFixes,
-    }, {
-      agentName: agent,
-      recreateOutputLabel: 'Fix with AI',
-      failureOutputLabel: 'Fix with AI',
-      reviewGateContext: parsed.reviewGateContext,
-      signal: deps.signal,
-    });
-    await finalizeMutationWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.fix-with-agent',
-      started: result.started,
-      mutationTiming: deps.mutationTiming,
-      ...(result.kind === 'recreateWorkflowFromFreshBase'
-        ? { scopedWorkflowId: result.workflowId }
-        : { scopedTaskIds: [taskId] }),
-    });
-    if (result.kind === 'recreateWorkflowFromFreshBase') {
-      process.stdout.write(
-        `Startup merge conflict detected; recreated workflow ${result.workflowId} from a fresh base.\n`,
-      );
-      return;
-    }
-    process.stdout.write(
-      result.autoApproved
-        ? `Fix applied and auto-approved for task: ${taskId} (${agent}).\n`
-        : `Fix applied for task: ${taskId} (${agent}). Use 'approve ${taskId}' or 'reject ${taskId}' to finalize.\n`,
-    );
-  } catch (err) {
-    await finalizeMutationWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.fix-with-agent.failure',
-      mutationTiming: deps.mutationTiming,
-    });
-    throw err;
-  } finally {
-    autoFix.unsubscribe();
-  }
-}
-
-async function headlessResolveConflict(taskId: string, deps: HeadlessDeps, agentArg?: string): Promise<void> {
-  if (!taskId) throw new Error('Missing taskId. Usage: --headless resolve-conflict <taskId> [claude|codex]');
-  const restored = restoreWorkflowForTaskUnlessDeleteAllWon(taskId, deps, 'resolve-conflict');
-  if (!restored) return;
-  taskId = restored.resolvedTaskId;
-
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  const agent = (agentArg ?? 'claude').toLowerCase();
-  try {
-    const result = await resolveConflictAction(taskId, {
-      ...deps,
-      taskExecutor: te,
-      autoApproveAIFixes: deps.invokerConfig.autoApproveAIFixes,
-    }, agent, deps.signal);
-    await finalizeMutationWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.resolve-conflict',
-      started: result.started,
-      mutationTiming: deps.mutationTiming,
-      scopedTaskIds: [taskId],
-    });
-    process.stdout.write(
-      deps.invokerConfig.autoApproveAIFixes
-        ? `Conflict resolved and auto-approved for task: ${taskId} (${agent}).\n`
-        : `Conflict resolved for task: ${taskId} (${agent}). Use 'approve ${taskId}' or 'reject ${taskId}' to finalize.\n`,
-    );
-  } catch (err) {
-    await finalizeMutationWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.resolve-conflict.failure',
-      mutationTiming: deps.mutationTiming,
-    });
-    throw err;
-  } finally {
-    autoFix.unsubscribe();
-  }
-}
-
-async function headlessRebaseRetry(target: string, deps: HeadlessDeps): Promise<void> {
-  if (!target) throw new Error('Missing arguments. Usage: --headless rebase-retry <workflowId|mergeTaskId|taskId>');
-  const workflowId = resolveHeadlessTargetWorkflowId(target, deps.persistence);
-  await preemptWorkflowBeforeMutation(workflowId, {
-    preemptWorkflowExecution: (id) => preemptWorkflowExecution(id, deps),
-    logger: deps.logger,
-    context: 'headless.rebase-retry',
-    mutationTiming: deps.mutationTiming,
-  });
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  const started = await rebaseRetry(target, {
-    ...deps,
-    logger: deps.logger,
-    commandService: deps.commandService,
-    taskExecutor: te,
-    mutationTiming: deps.mutationTiming,
-  });
-  const runnable = started.filter(isDispatchableLaunch);
-  const { topup } = await dispatchStartedTasksWithGlobalTopup({
-    orchestrator: deps.orchestrator,
-    taskExecutor: te,
-    logger: deps.logger,
-    context: 'headless.rebase-retry',
-    started,
-    scopedWorkflowId: workflowId,
-    mutationTiming: deps.mutationTiming,
-  });
-  if (runnable.length + topup.length === 0) {
-    autoFix.unsubscribe();
-    return;
-  }
-  if (deps.noTrack) {
-    process.stdout.write('[headless] --no-track enabled: rebase-retry accepted; exiting without tracking.\n');
-    autoFix.unsubscribe();
-    return;
-  }
-  await trackHeadlessWorkflow(workflowId, deps, {
-    hasBackgroundWork: autoFix.isBusy,
-    printSummary: false,
-    printTaskOutput: true,
-    setExitCodeOnFailure: false,
-  });
-  autoFix.unsubscribe();
-
-  const tasksStarted = runnable.length;
-  process.stdout.write(`Rebase-retry: retried workflow from fresh base (${tasksStarted} task(s))\n`);
-}
-
-async function headlessRebaseRecreate(workflowTarget: string, deps: HeadlessDeps): Promise<void> {
-  if (!workflowTarget) throw new Error('Missing arguments. Usage: --headless rebase-recreate <workflowId|mergeTaskId|taskId>');
-  const workflowId = resolveHeadlessTargetWorkflowId(workflowTarget, deps.persistence);
-  await preemptWorkflowBeforeMutation(workflowId, {
-    preemptWorkflowExecution: (id) => preemptWorkflowExecution(id, deps),
-    logger: deps.logger,
-    context: 'headless.rebase-recreate',
-    mutationTiming: deps.mutationTiming,
-  });
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  const started = await rebaseRecreate(workflowTarget, {
-    ...deps,
-    logger: deps.logger,
-    commandService: deps.commandService,
-    taskExecutor: te,
-    mutationTiming: deps.mutationTiming,
-  });
-  const runnable = started.filter(isDispatchableLaunch);
-  const { topup } = await dispatchStartedTasksWithGlobalTopup({
-    orchestrator: deps.orchestrator,
-    taskExecutor: te,
-    logger: deps.logger,
-    context: 'headless.rebase-recreate',
-    started,
-    scopedWorkflowId: workflowId,
-    mutationTiming: deps.mutationTiming,
-  });
-  if (runnable.length + topup.length === 0) {
-    autoFix.unsubscribe();
-    return;
-  }
-  if (deps.noTrack) {
-    process.stdout.write('[headless] --no-track enabled: rebase-recreate accepted; exiting without tracking.\n');
-    autoFix.unsubscribe();
-    return;
-  }
-  await trackHeadlessWorkflow(workflowId, deps, {
-    hasBackgroundWork: autoFix.isBusy,
-    printSummary: false,
-    printTaskOutput: true,
-    setExitCodeOnFailure: false,
-  });
-  autoFix.unsubscribe();
-
-  const tasksStarted = runnable.length;
-  process.stdout.write(`Rebase-recreate: recreated workflow from fresh base (${tasksStarted} task(s))\n`);
-}
-
-async function headlessRecreateWorkflow(workflowId: string, deps: HeadlessDeps): Promise<void> {
-  if (!workflowId) {
-    throw new Error('Missing arguments. Usage: --headless recreate <workflowId>');
-  }
-  await preemptWorkflowBeforeMutation(workflowId, {
-    preemptWorkflowExecution: (id) => preemptWorkflowExecution(id, deps),
-    logger: deps.logger,
-    context: 'headless.recreate-workflow',
-    mutationTiming: deps.mutationTiming,
-  });
-  const recreateWfEnvelope = makeEnvelope('recreate-workflow', 'headless', 'workflow', { workflowId });
-  const recreateWfResult = deps.mutationTiming
-    ? await deps.mutationTiming.span(
-      'headless.recreate-workflow.commandService.recreateWorkflow',
-      undefined,
-      () => deps.commandService.recreateWorkflow(recreateWfEnvelope),
-    )
-    : await deps.commandService.recreateWorkflow(recreateWfEnvelope);
-  if (!recreateWfResult.ok) throw new Error(recreateWfResult.error.message);
-  const started = recreateWfResult.data;
-  const runnable = started.filter(isDispatchableLaunch);
-  if (runnable.length > 0) {
-    const te = createHeadlessExecutor(deps);
-    const autoFix = wireHeadlessAutoFix(deps, te);
-    remoteFetchForPool.enabled = false;
-    let topup: TaskState[] = [];
-    try {
-      await te.executeTasks(runnable);
-      topup = await executeGlobalTopup({
-        orchestrator: deps.orchestrator,
-        taskExecutor: te,
-        logger: deps.logger,
-        context: 'headless.recreate-workflow',
-        alreadyDispatched: runnable,
-        mutationTiming: deps.mutationTiming,
-      });
-    } finally {
-      remoteFetchForPool.enabled = true;
-    }
-    if (runnable.length + topup.length === 0) {
-      autoFix.unsubscribe();
-      return;
-    }
-    if (deps.noTrack) {
-      process.stdout.write('[headless] --no-track enabled: recreate accepted; exiting without tracking.\n');
-      autoFix.unsubscribe();
-      return;
-    }
-    await trackHeadlessWorkflow(workflowId, deps, {
-      hasBackgroundWork: autoFix.isBusy,
-      printSummary: false,
-      printTaskOutput: true,
-      setExitCodeOnFailure: false,
-    });
-    autoFix.unsubscribe();
-  }
-  const tasksStarted = runnable.length;
-  process.stdout.write(`Recreate workflow "${workflowId}" — ${tasksStarted} task(s) to execute (pool fetch skipped)\n`);
-}
-
-async function headlessRecreateTask(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId) {
-    throw new Error('Missing arguments. Usage: --headless recreate-task <taskId>');
-  }
-  const restored = restoreWorkflowForTaskUnlessDeleteAllWon(taskId, deps, 'recreate-task');
-  if (!restored) return;
-  taskId = restored.resolvedTaskId;
-  if (deps.mutationTiming) {
-    await deps.mutationTiming.span(
-      'headless.recreate-task.preemptTaskSubgraph',
-      { taskId },
-      () => preemptTaskSubgraph(taskId, deps),
-    );
-  } else {
-    await preemptTaskSubgraph(taskId, deps);
-  }
-
-  const recreateTaskEnvelope = makeEnvelope('recreate-task', 'headless', 'task', { taskId });
-  const recreateTaskResult = deps.mutationTiming
-    ? await deps.mutationTiming.span(
-      'headless.recreate-task.commandService.recreateTask',
-      { taskId },
-      () => deps.commandService.recreateTask(recreateTaskEnvelope),
-    )
-    : await deps.commandService.recreateTask(recreateTaskEnvelope);
-  if (!recreateTaskResult.ok) throw new Error(recreateTaskResult.error.message);
-  const started = recreateTaskResult.data;
-  const runnable = started.filter(isDispatchableLaunch);
-  const workflowId = deps.orchestrator.getTask(taskId)?.config.workflowId;
-  process.stdout.write(`Recreate task "${taskId}" (+ downstream) — ${runnable.length} task(s) to execute (pool fetch skipped)\n`);
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  remoteFetchForPool.enabled = false;
-  let topup: TaskState[] = [];
-  try {
-    ({ topup } = await dispatchStartedTasksWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.recreate-task',
-      started,
-      mutationTiming: deps.mutationTiming,
-    }));
-  } finally {
-    remoteFetchForPool.enabled = true;
-  }
-  if (runnable.length + topup.length === 0) {
-    autoFix.unsubscribe();
-    return;
-  }
-  if (deps.noTrack) {
-    process.stdout.write('[headless] --no-track enabled: recreate-task accepted; exiting without tracking.\n');
-    autoFix.unsubscribe();
-    return;
-  }
-  if (workflowId) {
-    await trackHeadlessWorkflow(workflowId, deps, {
-      hasBackgroundWork: autoFix.isBusy,
-      printSummary: false,
-      printTaskOutput: true,
-      setExitCodeOnFailure: false,
-    });
-  }
-  autoFix.unsubscribe();
-}
-
-async function headlessRecreateDownstream(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId) {
-    throw new Error('Missing arguments. Usage: --headless recreate-downstream <taskId>');
-  }
-  const restored = restoreWorkflowForTaskUnlessDeleteAllWon(taskId, deps, 'recreate-downstream');
-  if (!restored) return;
-  taskId = restored.resolvedTaskId;
-  if (deps.mutationTiming) {
-    await deps.mutationTiming.span(
-      'headless.recreate-downstream.preemptTaskSubgraph',
-      { taskId },
-      () => preemptTaskSubgraph(taskId, deps),
-    );
-  } else {
-    await preemptTaskSubgraph(taskId, deps);
-  }
-
-  const envelope = makeEnvelope('recreate-downstream', 'headless', 'task', { taskId });
-  const result = deps.mutationTiming
-    ? await deps.mutationTiming.span(
-      'headless.recreate-downstream.commandService.recreateDownstream',
-      { taskId },
-      () => deps.commandService.recreateDownstream(envelope),
-    )
-    : await deps.commandService.recreateDownstream(envelope);
-  if (!result.ok) throw new Error(result.error.message);
-  const started = result.data;
-  const runnable = started.filter(isDispatchableLaunch);
-  const workflowId = deps.orchestrator.getTask(taskId)?.config.workflowId;
-  process.stdout.write(`Recreate downstream of "${taskId}" — ${runnable.length} task(s) to execute (pool fetch skipped)\n`);
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  remoteFetchForPool.enabled = false;
-  let topup: TaskState[] = [];
-  try {
-    ({ topup } = await dispatchStartedTasksWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.recreate-downstream',
-      started,
-      mutationTiming: deps.mutationTiming,
-    }));
-  } finally {
-    remoteFetchForPool.enabled = true;
-  }
-  if (runnable.length + topup.length === 0) {
-    autoFix.unsubscribe();
-    return;
-  }
-  if (deps.noTrack) {
-    process.stdout.write('[headless] --no-track enabled: recreate-downstream accepted; exiting without tracking.\n');
-    autoFix.unsubscribe();
-    return;
-  }
-  if (workflowId) {
-    await trackHeadlessWorkflow(workflowId, deps, {
-      hasBackgroundWork: autoFix.isBusy,
-      printSummary: false,
-      printTaskOutput: true,
-      setExitCodeOnFailure: false,
-    });
-  }
-  autoFix.unsubscribe();
-}
-
-async function headlessForkWorkflow(
-  workflowId: string,
-  deps: HeadlessDeps,
-): Promise<void> {
-  if (!workflowId) {
-    throw new Error('Missing arguments. Usage: --headless fork-workflow <workflowId>');
-  }
-  const result = sharedForkWorkflow(workflowId, {
-    orchestrator: deps.orchestrator,
-    logger: deps.logger,
-  });
-  const taskExecutor = createHeadlessExecutor(deps);
-  const { runnable } = await dispatchStartedTasksWithGlobalTopup({
-    orchestrator: deps.orchestrator,
-    taskExecutor,
-    logger: deps.logger,
-    context: 'headless.fork-workflow',
-    started: result.started,
-    scopedWorkflowId: result.forkedWorkflowId,
-  });
-  process.stdout.write(
-    `Forked workflow ${result.sourceWorkflowId} → ${result.forkedWorkflowId}; ` +
-      `launched ${runnable.length} task(s)\n`,
-  );
-}
-
-async function headlessRetryWorkflow(workflowId: string, deps: HeadlessDeps): Promise<void> {
-  if (!workflowId) {
-    throw new Error('Missing arguments. Usage: --headless retry <workflowId>');
-  }
-  deps.logger.info(`headlessRetryWorkflow begin workflow="${workflowId}" noTrack=${deps.noTrack ? 'true' : 'false'}`, {
-    module: 'headless',
-  });
-  await preemptWorkflowBeforeMutation(workflowId, {
-    preemptWorkflowExecution: (id) => preemptWorkflowExecution(id, deps),
-    logger: deps.logger,
-    context: 'headless.retry-workflow',
-    mutationTiming: deps.mutationTiming,
-  });
-  const envelope = makeEnvelope('retry-workflow', 'headless', 'workflow', { workflowId });
-  const result = deps.mutationTiming
-    ? await deps.mutationTiming.span(
-      'headless.retry-workflow.commandService.retryWorkflow',
-      undefined,
-      () => deps.commandService.retryWorkflow(envelope),
-    )
-    : await deps.commandService.retryWorkflow(envelope);
-  if (!result.ok) throw new Error(result.error.message);
-  const statusCounts = result.data.reduce<Record<string, number>>((acc, task) => {
-    acc[task.status] = (acc[task.status] ?? 0) + 1;
-    return acc;
-  }, {});
-  deps.logger.info(
-    `headlessRetryWorkflow trace workflow="${workflowId}" retryResult total=${result.data.length} statusCounts=${JSON.stringify(statusCounts)}`,
-    { module: 'headless' },
-  );
-  const retryRunningSummary = result.data
-    .filter(isDispatchableLaunch)
-    .map((task) => `${task.id}(${task.config.workflowId ?? 'unknown'})`);
-  if (retryRunningSummary.length > 0) {
-    deps.logger.info(
-      `headlessRetryWorkflow trace workflow="${workflowId}" retryResult running=[${retryRunningSummary.join(', ')}]`,
-      { module: 'headless' },
-    );
-  }
-  const runnable = result.data.filter(isDispatchableLaunch);
-  const crossWorkflow = runnable.filter((t) => t.config.workflowId !== workflowId);
-  if (crossWorkflow.length > 0) {
-    deps.logger.info(
-      `headlessRetryWorkflow dispatching cross-workflow runnable tasks for "${workflowId}": ${crossWorkflow.map((task) => `${task.id}(${task.config.workflowId ?? 'unknown'})`).join(', ')}`,
-      { module: 'headless' },
-    );
-  }
-  deps.logger.info(`headlessRetryWorkflow retry complete workflow="${workflowId}" runnable=${runnable.length}`, {
-    module: 'headless',
-  });
-
-  const runningKey = (task: TaskState): string => {
-    const attemptId = task.execution.selectedAttemptId?.trim();
-    return attemptId ? `attempt:${attemptId}` : `task:${task.id}`;
-  };
-  const scopedKeys = new Set(runnable.map((task) => runningKey(task)));
-  const globalTopup = deps.orchestrator
-    .startExecution()
-    .filter(isDispatchableLaunch)
-    .filter((task) => !scopedKeys.has(runningKey(task)));
-  deps.logger.info(
-    `headlessRetryWorkflow trace workflow="${workflowId}" postStartExecution globalTopup=${globalTopup.length}`,
-    { module: 'headless' },
-  );
-  if (globalTopup.length > 0) {
-    deps.logger.info(
-      `headlessRetryWorkflow trace workflow="${workflowId}" globalTopup running=[${globalTopup.map((task) => `${task.id}(${task.config.workflowId ?? 'unknown'})`).join(', ')}]`,
-      { module: 'headless' },
-    );
-  }
-  const dispatchable = [...runnable, ...globalTopup];
-  deps.logger.info(
-    `headlessRetryWorkflow trace workflow="${workflowId}" dispatchable=${dispatchable.length} ids=[${dispatchable.map((task) => `${task.id}(${task.config.workflowId ?? 'unknown'})`).join(', ')}]`,
-    { module: 'headless' },
-  );
-
-  process.stdout.write(`Retry workflow "${workflowId}" — ${dispatchable.length} task(s) to execute (completed tasks preserved)\n`);
-  if (dispatchable.length === 0) return;
-
-  if (deps.noTrack) {
-    if (deps.deferRunnableTasks) {
-      deps.deferRunnableTasks(dispatchable, workflowId);
-    } else {
-      const te = createHeadlessExecutor(deps);
-      const launch = setTimeout(() => {
-        void te.executeTasks(dispatchable).catch((err) => {
-          deps.logger.error(
-            `background no-track workflow retry failed for ${workflowId}: ${err instanceof Error ? err.stack ?? err.message : String(err)}`,
-            { module: 'headless' },
-          );
-        });
-      }, 25);
-      launch.unref?.();
-    }
-    process.stdout.write('[headless] --no-track enabled: retry accepted; exiting without tracking.\n');
-    return;
-  }
-
-  const te = createHeadlessExecutor(deps);
-  const autoFix = wireHeadlessAutoFix(deps, te);
-  remoteFetchForPool.enabled = false;
-  let topup: TaskState[] = [];
-  try {
-    ({ topup } = await dispatchStartedTasksWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.retry-workflow',
-      started: dispatchable,
-      scopedWorkflowId: workflowId,
-      mutationTiming: deps.mutationTiming,
-    }));
-  } finally {
-    remoteFetchForPool.enabled = true;
-  }
-  if (runnable.length + topup.length === 0) {
-    autoFix.unsubscribe();
-    return;
-  }
-  await trackHeadlessWorkflow(workflowId, deps, {
-    hasBackgroundWork: autoFix.isBusy,
-    printSummary: false,
-    printTaskOutput: true,
-    setExitCodeOnFailure: false,
-  });
-  autoFix.unsubscribe();
-}
-
-/** Orchestrator error codes that preemption treats as benign (cancel is best-effort). */
-const preemptSkipCodes: ReadonlySet<string> = new Set([
-  OrchestratorErrorCode.TASK_NOT_FOUND,
-  OrchestratorErrorCode.TASK_ALREADY_TERMINAL,
-  OrchestratorErrorCode.WORKFLOW_NOT_FOUND,
-]);
-
-async function preemptTaskSubgraph(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (deps.preemptTaskSubgraph) {
-    await deps.preemptTaskSubgraph(taskId);
-    return;
-  }
-  if (typeof deps.commandService.cancelTask !== 'function') return;
-  const envelope = makeEnvelope('cancel-task', 'headless', 'task', { taskId });
-  const result = await deps.commandService.cancelTask(envelope);
-  if (!result.ok) {
-    if (preemptSkipCodes.has(result.error.code)) return;
-    throw new Error(result.error.message);
-  }
-}
-
-async function preemptWorkflowExecution(workflowId: string, deps: HeadlessDeps): Promise<WorkflowCancelResult> {
-  if (deps.preemptWorkflowExecution) {
-    return deps.preemptWorkflowExecution(workflowId);
-  }
-  if (typeof deps.commandService.cancelWorkflow !== 'function') {
-    return { cancelled: [], runningCancelled: [] };
-  }
-  const envelope = makeEnvelope('cancel-workflow', 'headless', 'workflow', { workflowId });
-  const result = await deps.commandService.cancelWorkflow(envelope);
-  if (!result.ok) {
-    if (preemptSkipCodes.has(result.error.code)) return { cancelled: [], runningCancelled: [] };
-    throw new Error(result.error.message);
-  }
-  return result.data;
-}
+// ── Configure: edit family ───────────────────────────────────
 
 async function headlessEdit(taskId: string, newCommand: string, deps: HeadlessDeps): Promise<void> {
   if (!taskId || !newCommand) throw new Error('Missing arguments. Usage: --headless edit <taskId> <newCommand>');
@@ -2478,308 +705,6 @@ async function headlessEditAgent(taskId: string, agentName: string, deps: Headle
     setExitCodeOnFailure: false,
   });
   autoFix.unsubscribe();
-}
-
-async function headlessQuerySelect(taskId: string, deps: Pick<HeadlessDeps, 'persistence'>): Promise<void> {
-  if (!taskId) throw new Error('Missing taskId.');
-  const selected = deps.persistence.getSelectedExperiment(taskId);
-  process.stdout.write((selected
-    ? `Selected experiment for ${taskId}: ${selected}`
-    : `No experiment selected for ${taskId}`) + '\n');
-}
-
-/**
- * Resolve an agent session by ID via registered SessionDriver.
- * Shared by IPC handler (main.ts) and headless CLI (below).
- *
- * Flow: driver.loadSession() → driver.fetchRemoteSession() → driver.parseSession().
- * Each agent owns its own session resolution logic.
- */
-export async function resolveAgentSession(
-  sessionId: string,
-  agentName: string,
-  registry?: AgentRegistry,
-  allTasks?: import('@invoker/workflow-core').TaskState[],
-): Promise<AgentSessionData | null> {
-  const driver = registry?.getSessionDriver(agentName);
-  if (!driver) {
-    return {
-      agentName,
-      sessionId,
-      state: 'error',
-      messages: [],
-      reason: `No session driver registered for agent "${agentName}"`,
-    };
-  }
-
-  // 1. Try local
-  const raw = driver.loadSession(sessionId);
-  if (raw) {
-    const inspection = driver.inspectSession(raw);
-    return {
-      agentName,
-      sessionId,
-      state: inspection.state,
-      reason: inspection.reason,
-      messages: driver.parseSession(raw),
-      source: 'local',
-    };
-  }
-
-  // 2. Try remote (SSH tasks)
-  if (driver.fetchRemoteSession && allTasks) {
-    const sshTask = allTasks.find(
-      t => t.execution.agentSessionId === sessionId
-        && t.config.runnerKind === 'ssh',
-    );
-    if (sshTask) {
-      const { loadConfig } = await import('./config.js');
-      const targets = loadConfig().remoteTargets ?? {};
-      const targetId = (sshTask.config as { poolMemberId?: string }).poolMemberId;
-      const target = targetId
-        ? targets[targetId]
-        : Object.values(targets)[0];
-      if (target) {
-        const remoteRaw = await driver.fetchRemoteSession(sessionId, target);
-        if (remoteRaw) {
-          const inspection = driver.inspectSession(remoteRaw);
-          return {
-            agentName,
-            sessionId,
-            state: inspection.state,
-            reason: inspection.reason,
-            messages: driver.parseSession(remoteRaw),
-            source: 'remote',
-          };
-        }
-      }
-    }
-  }
-
-  return {
-    agentName,
-    sessionId,
-    state: 'error',
-    messages: [],
-    reason: 'Session file not found',
-  };
-}
-
-async function headlessSession(taskId: string | undefined, deps: Pick<HeadlessDeps, 'orchestrator' | 'persistence' | 'executionAgentRegistry'>): Promise<void> {
-  if (!taskId) throw new Error('Usage: --headless session <taskId>');
-  taskId = restoreWorkflowForTask(taskId, deps).resolvedTaskId;
-  const task = deps.orchestrator.getTask(taskId);
-  if (!task) throw new Error(`Task "${taskId}" not found`);
-
-  let sessionId = task.execution.agentSessionId ?? task.execution.lastAgentSessionId;
-  let agentName = task.execution.agentName ?? task.execution.lastAgentName ?? 'claude';
-
-  // Fallback: if current execution dropped agentSessionId, recover the most
-  // recent session from task event payloads.
-  if (!sessionId) {
-    const events = deps.persistence.getEvents(taskId) ?? [];
-    for (let i = events.length - 1; i >= 0; i -= 1) {
-      const payload = events[i].payload;
-      if (!payload) continue;
-      try {
-        const parsed = JSON.parse(payload);
-        const exec = parsed?.execution;
-        if (exec?.agentSessionId) {
-          sessionId = String(exec.agentSessionId);
-          if (exec.agentName) {
-            agentName = String(exec.agentName);
-          }
-          process.stdout.write(`Recovered agent session from event log: ${sessionId}\n`);
-          break;
-        }
-      } catch {
-        // Ignore malformed payload JSON
-      }
-    }
-  }
-
-  if (!sessionId) {
-    process.stdout.write(`No agent session for task "${taskId}"\n`);
-    return;
-  }
-
-  process.stdout.write(`agent=${agentName} sessionId=${sessionId}\n`);
-
-  const allTasks = deps.orchestrator.getAllTasks();
-  const result = await resolveAgentSession(sessionId, agentName, deps.executionAgentRegistry, allTasks);
-  if (!result) {
-    process.stdout.write('Session lookup failed\n');
-    return;
-  }
-  process.stdout.write(`state=${result.state}${result.source ? ` source=${result.source}` : ''}\n`);
-  if (result.reason) {
-    process.stdout.write(`${result.reason}\n`);
-  }
-  for (const msg of result.messages) {
-    process.stdout.write(`[${msg.role}] ${msg.content}\n`);
-  }
-}
-
-async function headlessCancel(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId) throw new Error('Missing taskId. Usage: --headless cancel <taskId>');
-  await withRestoredTaskUnlessDeleteAllWon(taskId, deps, 'cancel', async (restored) => {
-    taskId = restored.resolvedTaskId;
-
-    if (deps.cancelTask) {
-      const result = await deps.cancelTask(taskId);
-      process.stdout.write(`Cancelled ${result.cancelled.length} task(s): [${result.cancelled.join(', ')}]\n`);
-      if (result.runningCancelled.length > 0) {
-        process.stdout.write(`Killed running: [${result.runningCancelled.join(', ')}]\n`);
-      }
-      return;
-    }
-
-    const port = process.env.INVOKER_API_PORT;
-    if (port) {
-      const url = `http://127.0.0.1:${port}/api/tasks/${encodeURIComponent(taskId)}/cancel`;
-      try {
-        const ac = new AbortController();
-        const timer = setTimeout(() => ac.abort(), 10_000);
-        const res = await fetch(url, { method: 'POST', signal: ac.signal });
-        clearTimeout(timer);
-        if (res.ok) {
-          const data = (await res.json()) as { cancelled?: string[]; runningCancelled?: string[] };
-          const cancelled = data.cancelled ?? [];
-          const runningCancelled = data.runningCancelled ?? [];
-          process.stdout.write(`Cancelled ${cancelled.length} task(s): [${cancelled.join(', ')}]\n`);
-          if (runningCancelled.length > 0) {
-            process.stdout.write(`Killed running: [${runningCancelled.join(', ')}]\n`);
-          }
-          return;
-        }
-      } catch {
-        /* API unreachable — fall back to DB-only cancel */
-      }
-    }
-
-    const envelope = makeEnvelope('cancel-task', 'headless', 'task', { taskId });
-    const cmdResult = await deps.commandService.cancelTask(envelope);
-    if (!cmdResult.ok) throw new Error(cmdResult.error.message);
-    const te = createHeadlessExecutor(deps);
-    await finalizeMutationWithGlobalTopup({
-      orchestrator: deps.orchestrator,
-      taskExecutor: te,
-      logger: deps.logger,
-      context: 'headless.cancel-task',
-      mutationTiming: deps.mutationTiming,
-    });
-    process.stdout.write(`Cancelled ${cmdResult.data.cancelled.length} task(s): [${cmdResult.data.cancelled.join(', ')}]\n`);
-    if (cmdResult.data.runningCancelled.length > 0) {
-      process.stdout.write(`Killed running: [${cmdResult.data.runningCancelled.join(', ')}]\n`);
-    }
-  });
-}
-
-async function headlessCancelWorkflow(workflowId: string, deps: HeadlessDeps): Promise<void> {
-  if (!workflowId) throw new Error('Missing workflowId. Usage: --headless cancel-workflow <workflowId>');
-
-  if (deps.cancelWorkflow) {
-    const result = await deps.cancelWorkflow(workflowId);
-    process.stdout.write(
-      `Cancelled ${result.cancelled.length} task(s) in workflow "${workflowId}": [${result.cancelled.join(', ')}]\n`,
-    );
-    if (result.runningCancelled.length > 0) {
-      process.stdout.write(`Killed running: [${result.runningCancelled.join(', ')}]\n`);
-    }
-    return;
-  }
-
-  const port = process.env.INVOKER_API_PORT;
-  if (port) {
-    const url = `http://127.0.0.1:${port}/api/workflows/${encodeURIComponent(workflowId)}/cancel`;
-    try {
-      const ac = new AbortController();
-      const timer = setTimeout(() => ac.abort(), 10_000);
-      const res = await fetch(url, { method: 'POST', signal: ac.signal });
-      clearTimeout(timer);
-      if (res.ok) {
-        const data = (await res.json()) as { cancelled?: string[]; runningCancelled?: string[] };
-        const cancelled = data.cancelled ?? [];
-        const runningCancelled = data.runningCancelled ?? [];
-        process.stdout.write(
-          `Cancelled ${cancelled.length} task(s) in workflow "${workflowId}": [${cancelled.join(', ')}]\n`,
-        );
-        if (runningCancelled.length > 0) {
-          process.stdout.write(`Killed running: [${runningCancelled.join(', ')}]\n`);
-        }
-        return;
-      }
-    } catch {
-      /* fall back */
-    }
-  }
-
-  const result = await preemptWorkflowExecution(workflowId, deps);
-  const te = createHeadlessExecutor(deps);
-  await finalizeMutationWithGlobalTopup({
-    orchestrator: deps.orchestrator,
-    taskExecutor: te,
-    logger: deps.logger,
-    context: 'headless.cancel-workflow',
-    mutationTiming: deps.mutationTiming,
-  });
-  process.stdout.write(`Cancelled ${result.cancelled.length} task(s) in workflow "${workflowId}": [${result.cancelled.join(', ')}]\n`);
-  if (result.runningCancelled.length > 0) {
-    process.stdout.write(`Killed running: [${result.runningCancelled.join(', ')}]\n`);
-  }
-}
-
-async function headlessOpenTerminal(taskId: string, deps: HeadlessDeps): Promise<void> {
-  if (!taskId) throw new Error('Missing taskId. Usage: --headless open-terminal <taskId>');
-  const result = await openExternalTerminalForTask({
-    taskId,
-    persistence: deps.persistence,
-    executorRegistry: deps.executorRegistry,
-    executionAgentRegistry: deps.executionAgentRegistry,
-    repoRoot: deps.repoRoot,
-    logger: deps.logger,
-    runningTaskReason: 'Task is still running. View output in logs.',
-  });
-  if (result.opened) {
-    process.stdout.write(`Opened terminal for task: ${taskId}\n`);
-  } else {
-    process.stderr.write(`Could not open terminal: ${result.reason}\n`);
-    process.exitCode = 1;
-  }
-}
-
-async function headlessDeleteWorkflow(workflowId: string, deps: HeadlessDeps): Promise<void> {
-  if (!workflowId) throw new Error('Missing workflowId. Usage: --headless delete-workflow <workflowId>');
-  // Preempt running tasks (kill processes + cancel) — matches owner-mode bridge contract
-  await preemptWorkflowExecution(workflowId, deps);
-  const taskExecutor = createHeadlessExecutor(deps);
-  await taskExecutor.closeWorkflowReview(workflowId);
-  // Serialized via CommandService: DB delete + memory clear + scheduler cleanup + removal deltas
-  const envelope = makeEnvelope('delete-workflow', 'headless', 'workflow', { workflowId });
-  const result = await deps.commandService.deleteWorkflow(envelope);
-  if (!result.ok) throw new Error(result.error.message);
-  process.stdout.write(`Deleted workflow: ${workflowId}\n`);
-}
-
-async function headlessDetachWorkflow(
-  workflowId: string,
-  upstreamWorkflowId: string,
-  deps: Pick<HeadlessDeps, 'commandService'>,
-): Promise<void> {
-  if (!workflowId || !upstreamWorkflowId) {
-    throw new Error(
-      'Missing arguments. Usage: --headless detach-workflow <workflowId> <upstreamWorkflowId>',
-    );
-  }
-  const envelope = makeEnvelope('detach-workflow', 'headless', 'workflow', {
-    workflowId,
-    upstreamWorkflowId,
-  });
-  const result = await deps.commandService.detachWorkflow(envelope);
-  if (!result.ok) throw new Error(result.error.message);
-  process.stdout.write(
-    `Detached workflow ${workflowId} from upstream workflow ${upstreamWorkflowId}\n`,
-  );
 }
 
 /**
@@ -3052,103 +977,3 @@ async function headlessSlack(deps: HeadlessDeps): Promise<void> {
     process.on('SIGTERM', shutdown);
   });
 }
-
-// ── Headless Helpers ─────────────────────────────────────────
-
-function restoreWorkflowForTask(
-  taskId: string,
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'persistence'>,
-): { workflowId: string; resolvedTaskId: string } {
-  const restored = tryRestoreWorkflowForTask(taskId, deps);
-  if (restored) {
-    return restored;
-  }
-  throw new Error(`Task "${taskId}" not found in any workflow`);
-}
-
-function tryRestoreWorkflowForTask(
-  taskId: string,
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'persistence'>,
-): { workflowId: string; resolvedTaskId: string } | null {
-  const { orchestrator, persistence } = deps;
-  const workflows = persistence.listWorkflows();
-  for (const wf of workflows) {
-    const tasks = persistence.loadTasks(wf.id);
-    const match = tasks.find(t => t.id === taskId || t.id.endsWith('/' + taskId));
-    if (match) {
-      // Keep lookup read-only: load graph state from DB without starting tasks.
-      orchestrator.syncFromDb(wf.id);
-      return { workflowId: wf.id, resolvedTaskId: match.id };
-    }
-  }
-  return null;
-}
-
-function restoreWorkflowForTaskUnlessDeleteAllWon(
-  taskId: string,
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'persistence'>,
-  commandLabel: string,
-): { workflowId: string; resolvedTaskId: string } | null {
-  const restored = tryRestoreWorkflowForTask(taskId, deps);
-  if (restored) {
-    return restored;
-  }
-  if (deps.persistence.listWorkflows().length === 0) {
-    process.stdout.write(`[headless] ${commandLabel} skipped: task "${taskId}" was removed by delete-all.\n`);
-    return null;
-  }
-  throw new Error(`Task "${taskId}" not found in any workflow`);
-}
-
-async function withRestoredTaskUnlessDeleteAllWon<T>(
-  taskId: string,
-  deps: Pick<HeadlessDeps, 'orchestrator' | 'persistence'>,
-  commandLabel: string,
-  run: (restored: { workflowId: string; resolvedTaskId: string }) => Promise<T>,
-): Promise<T | undefined> {
-  const restored = restoreWorkflowForTaskUnlessDeleteAllWon(taskId, deps, commandLabel);
-  if (!restored) return undefined;
-  return await run(restored);
-}
-
-async function waitForCompletion(
-  orchestrator: Orchestrator,
-  workflowId?: string,
-  waitForApproval?: boolean,
-  hasBackgroundWork?: () => boolean,
-): Promise<void> {
-  const maxWaitMs = waitForApproval ? 86_400_000 : 1_800_000; // 24 hours if waiting for approval, else 30 minutes
-  const pollIntervalMs = 100;
-  const start = Date.now();
-
-  if (waitForApproval) {
-    process.stdout.write('[headless] Waiting for PR approval (--wait-for-approval)...\n');
-  }
-
-  while (Date.now() - start < maxWaitMs) {
-    let tasks = orchestrator.getAllTasks();
-    if (workflowId) {
-      tasks = tasks.filter((t) => t.config.workflowId === workflowId);
-    }
-    let readyTasks = orchestrator.getReadyTasks();
-    if (workflowId) {
-      readyTasks = readyTasks.filter((t) => t.config.workflowId === workflowId);
-    }
-    const settledStatuses = waitForApproval
-      ? ['completed', 'failed', 'closed', 'needs_input', 'blocked', 'stale']
-      : ['completed', 'failed', 'closed', 'needs_input', 'awaiting_approval', 'review_ready', 'blocked', 'stale'];
-    const allSettled = tasks.every((t) => settledStatuses.includes(t.status));
-    if (allSettled && !hasBackgroundWork?.()) return;
-    // Also settle if nothing is running and at least one task awaits human action.
-    // Pending merge gates can't progress until their upstream is approved.
-    const noneRunning = !tasks.some(
-      (t) => t.status === 'running' || t.status === 'fixing_with_ai',
-    );
-    const hasReadyPending = readyTasks.some((t) => t.status === 'pending');
-    const hasHumanBlocked = tasks.some((t) => settledStatuses.includes(t.status) && t.status !== 'completed');
-    if (noneRunning && hasHumanBlocked && !hasBackgroundWork?.()) return;
-    if (noneRunning && !hasReadyPending && !hasBackgroundWork?.()) return;
-    await new Promise((resolve) => setTimeout(resolve, pollIntervalMs));
-  }
-}
-// ── Headless Delegation ──────────────────────────────────────
